+++ com.unity.render-pipelines.high-definition/Runtime/RenderPipeline/HDRenderPipelineAsset.Migration.cs
@@ -11,7 +11,8 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             First,
             UpgradeFrameSettingsToStruct,
             AddAfterPostProcessFrameSetting,
-            AddFrameSettingSpecularLighting = 5
+            AddFrameSettingSpecularLighting = 5, // Not used anymore - don't removed the number
+            AddReflectionSettings
         }
 
         static readonly MigrationDescription<Version, HDRenderPipelineAsset> k_Migration = MigrationDescription.New(
@@ -31,11 +32,11 @@ namespace UnityEngine.Experimental.Rendering.HDPipeline
             {
                 FrameSettings.MigrateToAfterPostprocess(ref data.m_RenderingPathDefaultCameraFrameSettings);
             }),
-            MigrationStep.New(Version.AddFrameSettingSpecularLighting, (HDRenderPipelineAsset data) =>
+            MigrationStep.New(Version.AddReflectionSettings, (HDRenderPipelineAsset data) =>
             {
-                FrameSettings.MigrateToSpecularLighting(ref data.m_RenderingPathDefaultCameraFrameSettings);
-                FrameSettings.MigrateToSpecularLighting(ref data.m_RenderingPathDefaultBakedOrCustomReflectionFrameSettings);
-                FrameSettings.MigrateToSpecularLighting(ref data.m_RenderingPathDefaultRealtimeReflectionFrameSettings);
+                FrameSettings.MigrateToDefaultReflectionSettings(ref data.m_RenderingPathDefaultCameraFrameSettings);
+                FrameSettings.MigrateToNoReflectionSettings(ref data.m_RenderingPathDefaultBakedOrCustomReflectionFrameSettings);
+                FrameSettings.MigrateToNoReflectionRealtimeSettings(ref data.m_RenderingPathDefaultRealtimeReflectionFrameSettings);
             })
         );